refactor to C++ interface of coreneuron
