+++ test/validation/mod2c_core/cpp/NapDA.cpp
@@ -15,14 +15,16 @@
 
 #include "coreneuron/scopmath_core/newton_struct.h"
 #include "coreneuron/nrnoc/md2redef.h"
+#include "coreneuron/nrnoc/register_mech.hpp"
 #if !NRNGPU
 #if !defined(DISABLE_HOC_EXP)
 #undef exp
 #define exp hoc_Exp
 #endif
 #endif
+ namespace coreneuron {
  
-#define _thread_present_ /**/ , _thread[0:4] , _slist1[0:2], _dlist1[0:2] , _slist2[0:2] 
+#define _thread_present_ /**/ , _thread[0:3] , _slist1[0:2], _dlist1[0:2] , _slist2[0:2] 
  
 #if defined(_OPENACC) && !defined(DISABLE_OPENACC)
 #include <openacc.h>
@@ -71,23 +73,22 @@
 #define _STRIDE _cntml_padded + _iml
 #endif
  
-#define nrn_init _nrn_init__Is
-#define nrn_cur _nrn_cur__Is
-#define _nrn_current _nrn_current__Is
-#define nrn_jacob _nrn_jacob__Is
-#define nrn_state _nrn_state__Is
-#define initmodel initmodel__Is
-#define _net_receive _net_receive__Is
-#define nrn_state_launcher nrn_state_Is_launcher
-#define nrn_cur_launcher nrn_cur_Is_launcher
-#define nrn_jacob_launcher nrn_jacob_Is_launcher 
-#define rates rates_Is 
-#define states states_Is 
+#define nrn_init _nrn_init__NapDA
+#define nrn_cur _nrn_cur__NapDA
+#define _nrn_current _nrn_current__NapDA
+#define nrn_jacob _nrn_jacob__NapDA
+#define nrn_state _nrn_state__NapDA
+#define initmodel initmodel__NapDA
+#define _net_receive _net_receive__NapDA
+#define nrn_state_launcher nrn_state_NapDA_launcher
+#define nrn_cur_launcher nrn_cur_NapDA_launcher
+#define nrn_jacob_launcher nrn_jacob_NapDA_launcher 
+#define states states_NapDA 
  
 #define _threadargscomma_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v,
-#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v,
+#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v,
 #define _threadargs_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v
-#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v
+#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v
  	/*SUPPRESS 761*/
 	/*SUPPRESS 762*/
 	/*SUPPRESS 763*/
@@ -96,18 +97,16 @@
  
 #define t _nt->_t
 #define dt _nt->_dt
-#define gsbar _p[0*_STRIDE]
-#define ica _p[1*_STRIDE]
-#define ics _p[2*_STRIDE]
+#define gNapbar _p[0*_STRIDE]
+#define ena _p[1*_STRIDE]
+#define gna _p[2*_STRIDE]
 #define m _p[3*_STRIDE]
-#define n _p[4*_STRIDE]
-#define Dm _p[5*_STRIDE]
-#define Dn _p[6*_STRIDE]
-#define cai _p[7*_STRIDE]
-#define ins _p[8*_STRIDE]
-#define lca _p[9*_STRIDE]
-#define _v_unused _p[10*_STRIDE]
-#define _g_unused _p[11*_STRIDE]
+#define h _p[4*_STRIDE]
+#define ina _p[5*_STRIDE]
+#define Dm _p[6*_STRIDE]
+#define Dh _p[7*_STRIDE]
+#define _v_unused _p[8*_STRIDE]
+#define _g_unused _p[9*_STRIDE]
  
 #ifndef NRN_PRCELLSTATE
 #define NRN_PRCELLSTATE 0
@@ -119,13 +118,8 @@
 #define _PRCELLSTATE_V /**/
 #define _PRCELLSTATE_G /**/
 #endif
-#define _ion_cai		_nt_data[_ppvar[0*_STRIDE]]
-#define _ion_ica	_nt_data[_ppvar[1*_STRIDE]]
-#define _ion_dicadv	_nt_data[_ppvar[2*_STRIDE]]
-#define _ion_ics	_nt_data[_ppvar[3*_STRIDE]]
-#define _ion_dicsdv	_nt_data[_ppvar[4*_STRIDE]]
-#define _ion_ins	_nt_data[_ppvar[5*_STRIDE]]
-#define _ion_dinsdv	_nt_data[_ppvar[6*_STRIDE]]
+#define _ion_ina	_nt_data[_ppvar[0*_STRIDE]]
+#define _ion_dinadv	_nt_data[_ppvar[1*_STRIDE]]
  
 #if MAC
 #if !defined(v)
@@ -135,19 +129,16 @@
 #define h _mlhh
 #endif
 #endif
- 
-#if defined(__cplusplus)
-extern "C" {
-#endif
  static int hoc_nrnpointerindex =  -1;
  static ThreadDatum* _extcall_thread;
  /* external NEURON variables */
  
 #if 0 /*BBCORE*/
  /* declaration of user functions */
- static void _hoc_alp(void);
- static void _hoc_bet(void);
- static void _hoc_rates(void);
+ static void _hoc_hbet(void);
+ static void _hoc_half(void);
+ static void _hoc_mbet(void);
+ static void _hoc_malf(void);
  
 #endif /*BBCORE*/
  static int _mechtype;
@@ -155,30 +146,24 @@ extern "C" {
 #if 0 /*BBCORE*/
  /* connect user functions to hoc names */
  static VoidFunc hoc_intfunc[] = {
- "setdata_Is", _hoc_setdata,
- "alp_Is", _hoc_alp,
- "bet_Is", _hoc_bet,
- "rates_Is", _hoc_rates,
+ "setdata_NapDA", _hoc_setdata,
+ "hbet_NapDA", _hoc_hbet,
+ "half_NapDA", _hoc_half,
+ "mbet_NapDA", _hoc_mbet,
+ "malf_NapDA", _hoc_malf,
  0, 0
 };
  
 #endif /*BBCORE*/
-#define alp alp_Is
-#define bet bet_Is
- inline double alp( _threadargsprotocomma_ double , double );
- inline double bet( _threadargsprotocomma_ double , double );
+#define hbet hbet_NapDA
+#define half half_NapDA
+#define mbet mbet_NapDA
+#define malf malf_NapDA
+ inline double hbet( _threadargsprotocomma_ double );
+ inline double half( _threadargsprotocomma_ double );
+ inline double mbet( _threadargsprotocomma_ double );
+ inline double malf( _threadargsprotocomma_ double );
  /* declare global and static user variables */
- static int _thread1data_inuse = 0;
-static double _thread1data[4];
-#define _gth 3
-#define mtau_Is _thread1data[0]
-#define mtau _thread[_gth]._pval[0]
-#define minf_Is _thread1data[1]
-#define minf _thread[_gth]._pval[1]
-#define ntau_Is _thread1data[2]
-#define ntau _thread[_gth]._pval[2]
-#define ninf_Is _thread1data[3]
-#define ninf _thread[_gth]._pval[3]
  
 static void _acc_globals_update() {
  }
@@ -186,28 +171,22 @@ static void _acc_globals_update() {
 #if 0 /*BBCORE*/
  /* some parameters have upper and lower limits */
  static HocParmLimits _hoc_parm_limits[] = {
- "gsbar_Is", 0, 1e+09,
+ "gNapbar_NapDA", 0, 1e+09,
  0,0,0
 };
  static HocParmUnits _hoc_parm_units[] = {
- "mtau_Is", "ms",
- "ntau_Is", "ms",
- "gsbar_Is", "S/cm2",
- "ica_Is", "mA/cm2",
- "ics_Is", "mA/cm2",
+ "gNapbar_NapDA", "mho/cm2",
+ "ena_NapDA", "mV",
+ "gna_NapDA", "mho/cm2",
  0,0
 };
  
 #endif /*BBCORE*/
  static double delta_t = 0.01;
+ static double h0 = 0;
  static double m0 = 0;
- static double n0 = 0;
  /* connect global user variables to hoc */
  static DoubScal hoc_scdoub[] = {
- "minf_Is", &minf_Is,
- "ninf_Is", &ninf_Is,
- "mtau_Is", &mtau_Is,
- "ntau_Is", &ntau_Is,
  0,0
 };
  static DoubVec hoc_vdoub[] = {
@@ -215,42 +194,33 @@ static void _acc_globals_update() {
 };
  static double _sav_indep;
  static void nrn_alloc(double*, Datum*, int);
-void nrn_init(_NrnThread*, _Memb_list*, int);
-void nrn_state(_NrnThread*, _Memb_list*, int);
- void nrn_cur(_NrnThread*, _Memb_list*, int);
+void nrn_init(NrnThread*, Memb_list*, int);
+void nrn_state(NrnThread*, Memb_list*, int);
+ void nrn_cur(NrnThread*, Memb_list*, int);
  /* connect range variables in _p that hoc is supposed to know about */
  static const char *_mechanism[] = {
  "6.2.0",
-"Is",
- "gsbar_Is",
+"NapDA",
+ "gNapbar_NapDA",
+ "ena_NapDA",
  0,
- "ica_Is",
- "ics_Is",
+ "gna_NapDA",
  0,
- "m_Is",
- "n_Is",
+ "m_NapDA",
+ "h_NapDA",
  0,
  0};
- static int _ca_type;
- static int _cs_type;
- static int _ns_type;
+ static int _na_type;
  
 static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
  
 #if 0 /*BBCORE*/
  	/*initialize range parameters*/
- 	gsbar = 5e-05;
- prop_ion = need_memb(_ca_sym);
- nrn_promote(prop_ion, 1, 0);
- 	_ppvar[0]._pval = &prop_ion->param[1]; /* cai */
- 	_ppvar[1]._pval = &prop_ion->param[3]; /* ica */
- 	_ppvar[2]._pval = &prop_ion->param[4]; /* _ion_dicadv */
- prop_ion = need_memb(_cs_sym);
- 	_ppvar[3]._pval = &prop_ion->param[3]; /* ics */
- 	_ppvar[4]._pval = &prop_ion->param[4]; /* _ion_dicsdv */
- prop_ion = need_memb(_ns_sym);
- 	_ppvar[5]._pval = &prop_ion->param[3]; /* ins */
- 	_ppvar[6]._pval = &prop_ion->param[4]; /* _ion_dinsdv */
+ 	gNapbar = 0.0022;
+ 	ena = 55;
+ prop_ion = need_memb(_na_sym);
+ 	_ppvar[0]._pval = &prop_ion->param[3]; /* ina */
+ 	_ppvar[1]._pval = &prop_ion->param[4]; /* _ion_dinadv */
  
 #endif /* BBCORE */
  
@@ -260,47 +230,36 @@ static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
  static void _thread_cleanup(ThreadDatum*);
  static void _update_ion_pointer(Datum*);
  
-#define _psize 12
-#define _ppsize 7
- void _is_reg() {
+#define _psize 10
+#define _ppsize 2
+ void _NapDA_reg() {
 	int _vectorized = 1;
   _initlists();
  _mechtype = nrn_get_mechtype(_mechanism[1]);
  if (_mechtype == -1) return;
  _nrn_layout_reg(_mechtype, LAYOUT);
- _ca_type = nrn_get_mechtype("ca_ion"); _cs_type = nrn_get_mechtype("cs_ion"); _ns_type = nrn_get_mechtype("ns_ion"); 
+ _na_type = nrn_get_mechtype("na_ion"); 
 #if 0 /*BBCORE*/
- 	ion_reg("ca", -10000.);
- 	ion_reg("cs", 2.0);
- 	ion_reg("ns", 2.0);
- 	_ca_sym = hoc_lookup("ca_ion");
- 	_cs_sym = hoc_lookup("cs_ion");
- 	_ns_sym = hoc_lookup("ns_ion");
+ 	ion_reg("na", -10000.);
+ 	_na_sym = hoc_lookup("na_ion");
  
 #endif /*BBCORE*/
- 	register_mech(_mechanism, nrn_alloc,nrn_cur, NULL, nrn_state, nrn_init, hoc_nrnpointerindex, 5);
-  _extcall_thread = (ThreadDatum*)ecalloc(4, sizeof(ThreadDatum));
+ 	register_mech(_mechanism, nrn_alloc,nrn_cur, NULL, nrn_state, nrn_init, hoc_nrnpointerindex, 4);
+  _extcall_thread = (ThreadDatum*)ecalloc(3, sizeof(ThreadDatum));
   _thread_mem_init(_extcall_thread);
-  _thread1data_inuse = 0;
      _nrn_thread_reg1(_mechtype, _thread_mem_init);
      _nrn_thread_reg0(_mechtype, _thread_cleanup);
   hoc_register_prop_size(_mechtype, _psize, _ppsize);
-  hoc_register_dparam_semantics(_mechtype, 0, "ca_ion");
-  hoc_register_dparam_semantics(_mechtype, 1, "ca_ion");
-  hoc_register_dparam_semantics(_mechtype, 2, "ca_ion");
-  hoc_register_dparam_semantics(_mechtype, 3, "cs_ion");
-  hoc_register_dparam_semantics(_mechtype, 4, "cs_ion");
-  hoc_register_dparam_semantics(_mechtype, 5, "ns_ion");
-  hoc_register_dparam_semantics(_mechtype, 6, "ns_ion");
+  hoc_register_dparam_semantics(_mechtype, 0, "na_ion");
+  hoc_register_dparam_semantics(_mechtype, 1, "na_ion");
  	hoc_register_var(hoc_scdoub, hoc_vdoub, NULL);
  }
-static char *modelname = "Cardiac L-type Calcium channel";
+static char *modelname = "";
 
 static int error;
 static int _ninits = 0;
 static int _match_recurse=1;
 static void _modl_cleanup(){ _match_recurse=1;}
-static int rates(_threadargsprotocomma_ double);
  
 #define _deriv1_advance _thread[0]._i
 #define _dith1 1
@@ -309,38 +268,36 @@ static int rates(_threadargsprotocomma_ double);
 static int _ode_spec1(_threadargsproto_);
 /*static int _ode_matsol1(_threadargsproto_);*/
  
-/* _derivimplicit_ states _Is */
+/* _derivimplicit_ states _NapDA */
 #ifndef INSIDE_NMODL
 #define INSIDE_NMODL
 #endif
 #include "_kinderiv.h"
- int _newton_states_Is(_threadargsproto_);
+ int _newton_states_NapDA(_threadargsproto_);
  
-#define _slist2 _slist2_Is
+#define _slist2 _slist2_NapDA
 int* _slist2;
 #pragma acc declare create(_slist2)
   
-#define _slist1 _slist1_Is
+#define _slist1 _slist1_NapDA
 int* _slist1;
 #pragma acc declare create(_slist1)
 
-#define _dlist1 _dlist1_Is
+#define _dlist1 _dlist1_NapDA
 int* _dlist1;
 #pragma acc declare create(_dlist1)
  extern int states(_threadargsproto_);
  
 /*CVODE*/
  static int _ode_spec1 (_threadargsproto_) {int _reset = 0; {
-   rates ( _threadargscomma_ v ) ;
-   Dm = ( minf - m ) / mtau ;
-   Dn = ( ninf - n ) / ntau ;
+   Dm = ( 1.0 - m ) * malf ( _threadargscomma_ v ) - m * mbet ( _threadargscomma_ v ) ;
+   Dh = ( ( 1.0 - h ) * half ( _threadargscomma_ v ) - h * hbet ( _threadargscomma_ v ) ) / 2.0 ;
    }
  return _reset;
 }
  static int _ode_matsol1 (_threadargsproto_) {
- rates ( _threadargscomma_ v ) ;
- Dm = Dm  / (1. - dt*( ( ( ( - 1.0 ) ) ) / mtau )) ;
- Dn = Dn  / (1. - dt*( ( ( ( - 1.0 ) ) ) / ntau )) ;
+ Dm = Dm  / (1. - dt*( (( ( - 1.0 ) ))*(malf ( _threadargscomma_ v )) - (1.0)*(mbet ( _threadargscomma_ v )) )) ;
+ Dh = Dh  / (1. - dt*( ( ( (( ( - 1.0 ) ))*(half ( _threadargscomma_ v )) - (1.0)*(hbet ( _threadargscomma_ v )) ) ) / 2.0 )) ;
  return 0;
 }
  /*END CVODE*/
@@ -350,20 +307,19 @@ int states (_threadargsproto_) {int _reset=0; int error = 0;
  double* _dlist2 = (double*)(_thread[_dith1]._pval) + (2*_cntml_padded);
  {int _id; for(_id=0; _id < 2; _id++) { _savstate1[_id*_STRIDE] = _p[_slist1[_id]*_STRIDE];}}
  #pragma acc routine(nrn_newton_thread) seq
-_reset = nrn_newton_thread((NewtonSpace*)_newtonspace1, 2,_slist2, _derivimplicit_states_Is, _dlist2,  _threadargs_);
+_reset = nrn_newton_thread((NewtonSpace*)_newtonspace1, 2,_slist2, _derivimplicit_states_NapDA, _dlist2,  _threadargs_);
  /*if(_reset) {abort_run(_reset);}*/ }
  
   return _reset;
 }
 
-int _newton_states_Is (_threadargsproto_) {  int _reset=0;
+int _newton_states_NapDA (_threadargsproto_) {  int _reset=0;
  { double* _savstate1 = (double*)_thread[_dith1]._pval;
  double* _dlist2 = (double*)(_thread[_dith1]._pval) + (2*_cntml_padded);
  int _counte = -1;
  {
-   rates ( _threadargscomma_ v ) ;
-   Dm = ( minf - m ) / mtau ;
-   Dn = ( ninf - n ) / ntau ;
+   Dm = ( 1.0 - m ) * malf ( _threadargscomma_ v ) - m * mbet ( _threadargscomma_ v ) ;
+   Dh = ( ( 1.0 - h ) * half ( _threadargscomma_ v ) - h * hbet ( _threadargscomma_ v ) ) / 2.0 ;
    {int _id; for(_id=0; _id < 2; _id++) {
 if (_deriv1_advance) {
  _dlist2[(++_counte)*_STRIDE] = _p[_dlist1[_id]*_STRIDE] - (_p[_slist1[_id]*_STRIDE] - _savstate1[_id*_STRIDE])/dt;
@@ -373,118 +329,127 @@ _dlist2[(++_counte)*_STRIDE] = _p[_slist1[_id]*_STRIDE] - _savstate1[_id*_STRIDE
   } }
  return _reset;}
  
-double alp ( _threadargsprotocomma_ double _lv , double _li ) {
-   double _lalp;
- if ( _li  == 0.0 ) {
-     _lalp = 0.095 * exp ( - 0.01 * ( _lv - 5.0 ) ) / ( exp ( - 0.072 * ( _lv - 5.0 ) ) + 1.0 ) ;
+double malf ( _threadargsprotocomma_ double _lv ) {
+   double _lmalf;
+ double _lva ;
+ _lva = _lv + 12.0 ;
+   if ( fabs ( _lva ) < 1e-04 ) {
+     _lmalf = - 0.2816 * ( - 9.3 - _lva * 0.5 ) ;
      }
-   else if ( _li  == 1.0 ) {
-     _lalp = 0.012 * exp ( - 0.008 * ( _lv + 28.0 ) ) / ( exp ( 0.15 * ( _lv + 28.0 ) ) + 1.0 ) ;
+   else {
+     _lmalf = - 0.2816 * _lva / ( - 1.0 + exp ( - _lva / 9.3 ) ) ;
      }
    
-return _lalp;
+return _lmalf;
  }
  
 #if 0 /*BBCORE*/
  
-static void _hoc_alp(void) {
+static void _hoc_malf(void) {
   double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
   _thread = _extcall_thread;
   _nt = nrn_threads;
- _r =  alp ( _threadargs_, *getarg(1) , *getarg(2) ;
+ _r =  malf ( _threadargs_, *getarg(1) ;
  hoc_retpushx(_r);
 }
  
 #endif /*BBCORE*/
  
-double bet ( _threadargsprotocomma_ double _lv , double _li ) {
-   double _lbet;
- if ( _li  == 0.0 ) {
-     _lbet = 0.07 * exp ( - 0.017 * ( _lv + 44.0 ) ) / ( exp ( 0.05 * ( _lv + 44.0 ) ) + 1.0 ) ;
+double mbet ( _threadargsprotocomma_ double _lv ) {
+   double _lmbet;
+ double _lvb ;
+ _lvb = _lv - 15.0 ;
+   if ( fabs ( _lvb ) < 1e-04 ) {
+     _lmbet = 0.2464 * ( 6.0 - _lvb * 0.5 ) ;
      }
-   else if ( _li  == 1.0 ) {
-     _lbet = 0.0065 * exp ( - 0.02 * ( _lv + 30.0 ) ) / ( exp ( - 0.2 * ( _lv + 30.0 ) ) + 1.0 ) ;
+   else {
+     _lmbet = 0.2464 * _lvb / ( - 1.0 + exp ( _lvb / 6.0 ) ) ;
      }
    
-return _lbet;
+return _lmbet;
  }
  
 #if 0 /*BBCORE*/
  
-static void _hoc_bet(void) {
+static void _hoc_mbet(void) {
   double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
   _thread = _extcall_thread;
   _nt = nrn_threads;
- _r =  bet ( _threadargs_, *getarg(1) , *getarg(2) ;
+ _r =  mbet ( _threadargs_, *getarg(1) ;
  hoc_retpushx(_r);
 }
  
 #endif /*BBCORE*/
  
-static int  rates ( _threadargsprotocomma_ double _lv ) {
-   double _la , _lb ;
- _la = alp ( _threadargscomma_ _lv , 0.0 ) ;
-   _lb = bet ( _threadargscomma_ _lv , 0.0 ) ;
-   mtau = 1.0 / ( _la + _lb ) ;
-   minf = _la / ( _la + _lb ) ;
-   _la = alp ( _threadargscomma_ _lv , 1.0 ) ;
-   _lb = bet ( _threadargscomma_ _lv , 1.0 ) ;
-   ntau = 1.0 / ( _la + _lb ) ;
-   ninf = _la / ( _la + _lb ) ;
-    return 0; }
+double half ( _threadargsprotocomma_ double _lv ) {
+   double _lhalf;
+ _lhalf = 2.8e-5 * ( exp ( - ( _lv + 42.8477 ) / 4.0248 ) ) ;
+   
+return _lhalf;
+ }
  
 #if 0 /*BBCORE*/
  
-static void _hoc_rates(void) {
+static void _hoc_half(void) {
   double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
   _thread = _extcall_thread;
   _nt = nrn_threads;
- _r = 1.;
- rates ( _threadargs_, *getarg(1) ;
+ _r =  half ( _threadargs_, *getarg(1) ;
+ hoc_retpushx(_r);
+}
+ 
+#endif /*BBCORE*/
+ 
+double hbet ( _threadargsprotocomma_ double _lv ) {
+   double _lhbet;
+ _lhbet = 0.02 / ( 1.0 + exp ( - ( _lv - 413.9284 ) / 148.2589 ) ) ;
+   
+return _lhbet;
+ }
+ 
+#if 0 /*BBCORE*/
+ 
+static void _hoc_hbet(void) {
+  double _r;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
+   if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
+  _thread = _extcall_thread;
+  _nt = nrn_threads;
+ _r =  hbet ( _threadargs_, *getarg(1) ;
  hoc_retpushx(_r);
 }
  
 #endif /*BBCORE*/
  
 static void _thread_mem_init(ThreadDatum* _thread) {
-   _thread[_dith1]._pval = NULL;  if (_thread1data_inuse) {_thread[_gth]._pval = (double*)ecalloc(4, sizeof(double));
- }else{
- _thread[_gth]._pval = _thread1data; _thread1data_inuse = 1;
- }
- }
+   _thread[_dith1]._pval = NULL; }
  
 static void _thread_cleanup(ThreadDatum* _thread) {
    free( _thread[_dith1]._pval);
-   nrn_destroy_newtonspace(_newtonspace1);
-  if (_thread[_gth]._pval == _thread1data) {
-   _thread1data_inuse = 0;
-  }else{
-   free((void*)_thread[_gth]._pval);
-  }
+   nrn_destroy_newtonspace((NewtonSpace*) _newtonspace1);
  }
  static void _update_ion_pointer(Datum* _ppvar) {
  }
 
 static void initmodel(_threadargsproto_) {
   int _i; double _save;{
+  h = h0;
   m = m0;
-  n = n0;
  {
-   rates ( _threadargscomma_ v ) ;
-   m = minf ;
-   n = ninf ;
+   m = malf ( _threadargscomma_ v ) / ( malf ( _threadargscomma_ v ) + mbet ( _threadargscomma_ v ) ) ;
+   h = half ( _threadargscomma_ v ) / ( half ( _threadargscomma_ v ) + hbet ( _threadargscomma_ v ) ) ;
    }
  
 }
 }
 
-void nrn_init(_NrnThread* _nt, _Memb_list* _ml, int _type){
+void nrn_init(NrnThread* _nt, Memb_list* _ml, int _type){
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double _v, v; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -540,9 +505,8 @@ for (;;) { /* help clang-format properly indent */
     _PRCELLSTATE_V
  v = _v;
  _PRCELLSTATE_V
-  cai = _ion_cai;
  initmodel(_threadargs_);
-   }
+ }
   }
   _deriv1_advance = 1;
   #ifdef _OPENACC
@@ -551,27 +515,22 @@ for (;;) { /* help clang-format properly indent */
 }
 
 static double _nrn_current(_threadargsproto_, double _v){double _current=0.;v=_v;{ {
-   double _lEs ;
- _lEs = - 82.3 - 13.0287 * log ( cai ) ;
-   ics = gsbar * m * n * ( v - _lEs ) ;
-   ica = ics ;
-   ins = - ics ;
+   gna = gNapbar * m * h ;
+   ina = gna * ( v - ena ) ;
    }
- _current += ica;
- _current += ics;
- _current += ins;
+ _current += ina;
 
 } return _current;
 }
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC)
-  void nrn_state_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_jacob_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_cur_launcher(_NrnThread*, _Memb_list*, int, int);
+  void nrn_state_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_jacob_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_cur_launcher(NrnThread*, Memb_list*, int, int);
 #endif
 
 
-void nrn_cur(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_cur(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 int* _ni; double _rhs, _g, _v, v; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -582,8 +541,8 @@ double * _vec_rhs = _nt->_actual_rhs;
 double * _vec_d = _nt->_actual_d;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_cur_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -607,23 +566,14 @@ for (;;) { /* help clang-format properly indent */
     int _nd_idx = _ni[_iml];
     _v = _vec_v[_nd_idx];
     _PRCELLSTATE_V
-  cai = _ion_cai;
  _g = _nrn_current(_threadargs_, _v + .001);
- 	{ double _dins;
- double _dics;
- double _dica;
-  _dica = ica;
-  _dics = ics;
-  _dins = ins;
+ 	{ double _dina;
+  _dina = ina;
  _rhs = _nrn_current(_threadargs_, _v);
-  _ion_dicadv += (_dica - ica)/.001 ;
-  _ion_dicsdv += (_dics - ics)/.001 ;
-  _ion_dinsdv += (_dins - ins)/.001 ;
+  _ion_dinadv += (_dina - ina)/.001 ;
  	}
  _g = (_g - _rhs)/.001;
-  _ion_ica += ica ;
-  _ion_ics += ics ;
-  _ion_ins += ins ;
+  _ion_ina += ina ;
  _PRCELLSTATE_G
 	_vec_rhs[_nd_idx] -= _rhs;
 	_vec_d[_nd_idx] += _g;
@@ -632,7 +582,7 @@ for (;;) { /* help clang-format properly indent */
  
 }
 
-void nrn_state(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_state(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double v, _v = 0.0; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -641,8 +591,8 @@ _cntml_padded = _ml->_nodecount_padded;
 _thread = _ml->_thread;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_state_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -668,13 +618,12 @@ for (;;) { /* help clang-format properly indent */
     _PRCELLSTATE_V
  v=_v;
 {
-  cai = _ion_cai;
  {  
-  #if !defined(_derivimplicit_states_Is)
-    #define _derivimplicit_states_Is 0
+  #if !defined(_derivimplicit_states_NapDA)
+    #define _derivimplicit_states_NapDA 0
   #endif
-  derivimplicit_thread(2, _slist1, _dlist1, _derivimplicit_states_Is, _threadargs_);
-  }   }}
+  derivimplicit_thread(2, _slist1, _dlist1, _derivimplicit_states_NapDA, _threadargs_);
+  } }}
 
 }
 
@@ -691,18 +640,15 @@ static void _initlists(){
  _slist1 = (int*)malloc(sizeof(int)*2);
  _dlist1 = (int*)malloc(sizeof(int)*2);
  _slist1[0] = &(m) - _p;  _dlist1[0] = &(Dm) - _p;
- _slist1[1] = &(n) - _p;  _dlist1[1] = &(Dn) - _p;
+ _slist1[1] = &(h) - _p;  _dlist1[1] = &(Dh) - _p;
  #pragma acc enter data copyin(_slist1[0:2])
  #pragma acc enter data copyin(_dlist1[0:2])
 
  _slist2 = (int*)malloc(sizeof(int)*2);
- _slist2[0] = &(m) - _p;
- _slist2[1] = &(n) - _p;
+ _slist2[0] = &(h) - _p;
+ _slist2[1] = &(m) - _p;
  #pragma acc enter data copyin(_slist2[0:2])
 
 _first = 0;
 }
-
-#if defined(__cplusplus)
-} /* extern "C" */
-#endif
+} // namespace coreneuron_lib
