+++ test/validation/mod2c_core/cpp/Nap_E.cpp
@@ -15,14 +15,16 @@
 
 #include "coreneuron/scopmath_core/newton_struct.h"
 #include "coreneuron/nrnoc/md2redef.h"
+#include "coreneuron/nrnoc/register_mech.hpp"
 #if !NRNGPU
 #if !defined(DISABLE_HOC_EXP)
 #undef exp
 #define exp hoc_Exp
 #endif
 #endif
+ namespace coreneuron {
  
-#define _thread_present_ /**/ , _slist1[0:1], _dlist1[0:1] 
+#define _thread_present_ /**/ , _slist1[0:2], _dlist1[0:2] 
  
 #if defined(_OPENACC) && !defined(DISABLE_OPENACC)
 #include <openacc.h>
@@ -71,23 +73,23 @@
 #define _STRIDE _cntml_padded + _iml
 #endif
  
-#define nrn_init _nrn_init__NaSm
-#define nrn_cur _nrn_cur__NaSm
-#define _nrn_current _nrn_current__NaSm
-#define nrn_jacob _nrn_jacob__NaSm
-#define nrn_state _nrn_state__NaSm
-#define initmodel initmodel__NaSm
-#define _net_receive _net_receive__NaSm
-#define nrn_state_launcher nrn_state_NaSm_launcher
-#define nrn_cur_launcher nrn_cur_NaSm_launcher
-#define nrn_jacob_launcher nrn_jacob_NaSm_launcher 
-#define rates rates_NaSm 
-#define states states_NaSm 
+#define nrn_init _nrn_init__Nap_E
+#define nrn_cur _nrn_cur__Nap_E
+#define _nrn_current _nrn_current__Nap_E
+#define nrn_jacob _nrn_jacob__Nap_E
+#define nrn_state _nrn_state__Nap_E
+#define initmodel initmodel__Nap_E
+#define _net_receive _net_receive__Nap_E
+#define nrn_state_launcher nrn_state_Nap_E_launcher
+#define nrn_cur_launcher nrn_cur_Nap_E_launcher
+#define nrn_jacob_launcher nrn_jacob_Nap_E_launcher 
+#define rates rates_Nap_E 
+#define states states_Nap_E 
  
 #define _threadargscomma_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v,
-#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v,
+#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v,
 #define _threadargs_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v
-#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v
+#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v
  	/*SUPPRESS 761*/
 	/*SUPPRESS 762*/
 	/*SUPPRESS 763*/
@@ -96,15 +98,26 @@
  
 #define t _nt->_t
 #define dt _nt->_dt
-#define gnasmbar _p[0*_STRIDE]
-#define minf _p[1*_STRIDE]
-#define mtau _p[2*_STRIDE]
-#define gnasm _p[3*_STRIDE]
-#define m _p[4*_STRIDE]
-#define Dm _p[5*_STRIDE]
-#define ina _p[6*_STRIDE]
-#define _v_unused _p[7*_STRIDE]
-#define _g_unused _p[8*_STRIDE]
+#define gNap_Ebar _p[0*_STRIDE]
+#define ina _p[1*_STRIDE]
+#define gNap_E _p[2*_STRIDE]
+#define m _p[3*_STRIDE]
+#define h _p[4*_STRIDE]
+#define ttxo _p[5*_STRIDE]
+#define ttxi _p[6*_STRIDE]
+#define ena _p[7*_STRIDE]
+#define mInf _p[8*_STRIDE]
+#define mTau _p[9*_STRIDE]
+#define mAlpha _p[10*_STRIDE]
+#define mBeta _p[11*_STRIDE]
+#define hInf _p[12*_STRIDE]
+#define hTau _p[13*_STRIDE]
+#define hAlpha _p[14*_STRIDE]
+#define hBeta _p[15*_STRIDE]
+#define Dm _p[16*_STRIDE]
+#define Dh _p[17*_STRIDE]
+#define _v_unused _p[18*_STRIDE]
+#define _g_unused _p[19*_STRIDE]
  
 #ifndef NRN_PRCELLSTATE
 #define NRN_PRCELLSTATE 0
@@ -116,8 +129,11 @@
 #define _PRCELLSTATE_V /**/
 #define _PRCELLSTATE_G /**/
 #endif
-#define _ion_ina	_nt_data[_ppvar[0*_STRIDE]]
-#define _ion_dinadv	_nt_data[_ppvar[1*_STRIDE]]
+#define _ion_ena		_nt_data[_ppvar[0*_STRIDE]]
+#define _ion_ina	_nt_data[_ppvar[1*_STRIDE]]
+#define _ion_dinadv	_nt_data[_ppvar[2*_STRIDE]]
+#define _ion_ttxo		_nt_data[_ppvar[3*_STRIDE]]
+#define _ion_ttxi		_nt_data[_ppvar[4*_STRIDE]]
  
 #if MAC
 #if !defined(v)
@@ -126,10 +142,6 @@
 #if !defined(h)
 #define h _mlhh
 #endif
-#endif
- 
-#if defined(__cplusplus)
-extern "C" {
 #endif
  static int hoc_nrnpointerindex =  -1;
  static ThreadDatum* _extcall_thread;
@@ -145,43 +157,15 @@ extern "C" {
 #if 0 /*BBCORE*/
  /* connect user functions to hoc names */
  static VoidFunc hoc_intfunc[] = {
- "setdata_NaSm", _hoc_setdata,
- "rates_NaSm", _hoc_rates,
+ "setdata_Nap_E", _hoc_setdata,
+ "rates_Nap_E", _hoc_rates,
  0, 0
 };
  
 #endif /*BBCORE*/
  /* declare global and static user variables */
-#define Etemp Etemp_NaSm
- double Etemp = 21;
- #pragma acc declare copyin (Etemp)
-#define Vtm Vtm_NaSm
- double Vtm = -33.5;
- #pragma acc declare copyin (Vtm)
-#define Vsm Vsm_NaSm
- double Vsm = -16;
- #pragma acc declare copyin (Vsm)
-#define ena ena_NaSm
- double ena = 40;
- #pragma acc declare copyin (ena)
-#define ktm ktm_NaSm
- double ktm = 26.3;
- #pragma acc declare copyin (ktm)
-#define ksm ksm_NaSm
- double ksm = 9.4;
- #pragma acc declare copyin (ksm)
-#define tom tom_NaSm
- double tom = 637.8;
- #pragma acc declare copyin (tom)
  
 static void _acc_globals_update() {
- #pragma acc update device (Etemp) if(nrn_threads->compute_gpu)
- #pragma acc update device (Vtm) if(nrn_threads->compute_gpu)
- #pragma acc update device (Vsm) if(nrn_threads->compute_gpu)
- #pragma acc update device (ena) if(nrn_threads->compute_gpu)
- #pragma acc update device (ktm) if(nrn_threads->compute_gpu)
- #pragma acc update device (ksm) if(nrn_threads->compute_gpu)
- #pragma acc update device (tom) if(nrn_threads->compute_gpu)
  }
  
 #if 0 /*BBCORE*/
@@ -190,23 +174,18 @@ static void _acc_globals_update() {
  0,0,0
 };
  static HocParmUnits _hoc_parm_units[] = {
- "ena_NaSm", "mV",
- "gnasmbar_NaSm", "mho/cm2",
+ "gNap_Ebar_Nap_E", "S/cm2",
+ "ina_Nap_E", "mA/cm2",
+ "gNap_E_Nap_E", "S/cm2",
  0,0
 };
  
 #endif /*BBCORE*/
- static double delta_t = 1;
+ static double delta_t = 0.01;
+ static double h0 = 0;
  static double m0 = 0;
  /* connect global user variables to hoc */
  static DoubScal hoc_scdoub[] = {
- "ena_NaSm", &ena_NaSm,
- "Etemp_NaSm", &Etemp_NaSm,
- "Vsm_NaSm", &Vsm_NaSm,
- "ksm_NaSm", &ksm_NaSm,
- "tom_NaSm", &tom_NaSm,
- "Vtm_NaSm", &Vtm_NaSm,
- "ktm_NaSm", &ktm_NaSm,
  0,0
 };
  static DoubVec hoc_vdoub[] = {
@@ -214,32 +193,39 @@ static void _acc_globals_update() {
 };
  static double _sav_indep;
  static void nrn_alloc(double*, Datum*, int);
-void nrn_init(_NrnThread*, _Memb_list*, int);
-void nrn_state(_NrnThread*, _Memb_list*, int);
- void nrn_cur(_NrnThread*, _Memb_list*, int);
+void nrn_init(NrnThread*, Memb_list*, int);
+void nrn_state(NrnThread*, Memb_list*, int);
+ void nrn_cur(NrnThread*, Memb_list*, int);
  /* connect range variables in _p that hoc is supposed to know about */
  static const char *_mechanism[] = {
  "6.2.0",
-"NaSm",
- "gnasmbar_NaSm",
+"Nap_E",
+ "gNap_Ebar_Nap_E",
  0,
- "minf_NaSm",
- "mtau_NaSm",
- "gnasm_NaSm",
+ "ina_Nap_E",
+ "gNap_E_Nap_E",
  0,
- "m_NaSm",
+ "m_Nap_E",
+ "h_Nap_E",
  0,
  0};
  static int _na_type;
+ static int _ttx_type;
  
 static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
  
 #if 0 /*BBCORE*/
  	/*initialize range parameters*/
- 	gnasmbar = 0.00011;
+ 	gNap_Ebar = 1e-05;
  prop_ion = need_memb(_na_sym);
- 	_ppvar[0]._pval = &prop_ion->param[3]; /* ina */
- 	_ppvar[1]._pval = &prop_ion->param[4]; /* _ion_dinadv */
+ nrn_promote(prop_ion, 0, 1);
+ 	_ppvar[0]._pval = &prop_ion->param[0]; /* ena */
+ 	_ppvar[1]._pval = &prop_ion->param[3]; /* ina */
+ 	_ppvar[2]._pval = &prop_ion->param[4]; /* _ion_dinadv */
+ prop_ion = need_memb(_ttx_sym);
+ nrn_promote(prop_ion, 1, 0);
+ 	_ppvar[3]._pval = &prop_ion->param[2]; /* ttxo */
+ 	_ppvar[4]._pval = &prop_ion->param[1]; /* ttxi */
  
 #endif /* BBCORE */
  
@@ -247,84 +233,128 @@ static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
  static void _initlists();
  static void _update_ion_pointer(Datum*);
  
-#define _psize 9
-#define _ppsize 2
- void _NaSm_reg() {
+#define _psize 20
+#define _ppsize 5
+ void _Nap_E_reg() {
 	int _vectorized = 1;
   _initlists();
  _mechtype = nrn_get_mechtype(_mechanism[1]);
  if (_mechtype == -1) return;
  _nrn_layout_reg(_mechtype, LAYOUT);
- _na_type = nrn_get_mechtype("na_ion"); 
+ _na_type = nrn_get_mechtype("na_ion"); _ttx_type = nrn_get_mechtype("ttx_ion"); 
 #if 0 /*BBCORE*/
  	ion_reg("na", -10000.);
+ 	ion_reg("ttx", 1.0);
  	_na_sym = hoc_lookup("na_ion");
+ 	_ttx_sym = hoc_lookup("ttx_ion");
  
 #endif /*BBCORE*/
  	register_mech(_mechanism, nrn_alloc,nrn_cur, NULL, nrn_state, nrn_init, hoc_nrnpointerindex, 1);
   hoc_register_prop_size(_mechtype, _psize, _ppsize);
   hoc_register_dparam_semantics(_mechtype, 0, "na_ion");
   hoc_register_dparam_semantics(_mechtype, 1, "na_ion");
+  hoc_register_dparam_semantics(_mechtype, 2, "na_ion");
+  hoc_register_dparam_semantics(_mechtype, 3, "ttx_ion");
+  hoc_register_dparam_semantics(_mechtype, 4, "ttx_ion");
  	hoc_register_var(hoc_scdoub, hoc_vdoub, NULL);
  }
-static char *modelname = "A slow Sodium current";
+static char *modelname = "";
 
 static int error;
 static int _ninits = 0;
 static int _match_recurse=1;
 static void _modl_cleanup(){ _match_recurse=1;}
-static int rates(_threadargsprotocomma_ double);
+static int rates(_threadargsproto_);
  
 static int _ode_spec1(_threadargsproto_);
 /*static int _ode_matsol1(_threadargsproto_);*/
  
-#define _slist1 _slist1_NaSm
+#define _slist1 _slist1_Nap_E
 int* _slist1;
 #pragma acc declare create(_slist1)
 
-#define _dlist1 _dlist1_NaSm
+#define _dlist1 _dlist1_Nap_E
 int* _dlist1;
 #pragma acc declare create(_dlist1)
  static inline int states(_threadargsproto_);
  
 /*CVODE*/
  static int _ode_spec1 (_threadargsproto_) {int _reset = 0; {
-   rates ( _threadargscomma_ v ) ;
-   Dm = ( minf - m ) / mtau ;
+   if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
+     mInf = 0.0 ;
+     mTau = 1e-12 ;
+     hInf = 1.0 ;
+     hTau = 1e-12 ;
+     }
+   else {
+     rates ( _threadargs_ ) ;
+     }
+   Dm = ( mInf - m ) / mTau ;
+   Dh = ( hInf - h ) / hTau ;
    }
  return _reset;
 }
  static int _ode_matsol1 (_threadargsproto_) {
- rates ( _threadargscomma_ v ) ;
- Dm = Dm  / (1. - dt*( ( ( ( - 1.0 ) ) ) / mtau )) ;
+ if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
+   mInf = 0.0 ;
+   mTau = 1e-12 ;
+   hInf = 1.0 ;
+   hTau = 1e-12 ;
+   }
+ else {
+   rates ( _threadargs_ ) ;
+   }
+ Dm = Dm  / (1. - dt*( ( ( ( - 1.0 ) ) ) / mTau )) ;
+ Dh = Dh  / (1. - dt*( ( ( ( - 1.0 ) ) ) / hTau )) ;
  return 0;
 }
  /*END CVODE*/
  static int states (_threadargsproto_) { {
-   rates ( _threadargscomma_ v ) ;
-    m = m + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / mtau)))*(- ( ( ( minf ) ) / mtau ) / ( ( ( ( - 1.0) ) ) / mtau ) - m) ;
+   if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
+     mInf = 0.0 ;
+     mTau = 1e-12 ;
+     hInf = 1.0 ;
+     hTau = 1e-12 ;
+     }
+   else {
+     rates ( _threadargs_ ) ;
+     }
+    m = m + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / mTau)))*(- ( ( ( mInf ) ) / mTau ) / ( ( ( ( - 1.0) ) ) / mTau ) - m) ;
+    h = h + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / hTau)))*(- ( ( ( hInf ) ) / hTau ) / ( ( ( ( - 1.0) ) ) / hTau ) - h) ;
    }
   return 0;
 }
  
-static int  rates ( _threadargsprotocomma_ double _lv ) {
-   double _lq10 , _ltadj ;
- _lq10 = 2.5 ;
-   _ltadj = pow( _lq10 , ( ( celsius - Etemp ) / 10.0 ) ) ;
-   minf = 1.0 / ( 1.0 + exp ( - ( _lv - Vsm ) / ksm ) ) ;
-   mtau = tom / ( exp ( - ( _lv - Vtm ) / ktm ) + exp ( ( _lv - Vtm ) / ktm ) ) / _ltadj ;
-    return 0; }
+static int  rates ( _threadargsproto_ ) {
+    mInf = 1.0 / ( 1.0 + exp ( ( v - - 52.6 ) / - 4.6 ) ) ;
+   if ( v  == - 25.0 ) {
+     v = v + 0.0001 ;
+     }
+   mAlpha = ( 0.182 * ( v - - 25.0 ) ) / ( 1.0 - ( exp ( - ( v - - 25.0 ) / 9.0 ) ) ) ;
+   mBeta = ( 0.124 * ( - v - 25.0 ) ) / ( 1.0 - ( exp ( - ( - v - 25.0 ) / 9.0 ) ) ) ;
+   mTau = 6.0 * 1.0 / ( mAlpha + mBeta ) ;
+   if ( v  == - 50.0 ) {
+     v = v + 0.0001 ;
+     }
+   hAlpha = ( 0.00001 * ( v - - 50.0 ) ) / ( 1.0 - ( exp ( - ( v - - 50.0 ) / 13.0 ) ) ) ;
+   if ( v  == - 75.0 ) {
+     v = v + 0.0001 ;
+     }
+   hBeta = ( 0.00001 * ( - v - 75.0 ) ) / ( 1.0 - ( exp ( - ( - v - 75.0 ) / 13.0 ) ) ) ;
+   hInf = 1.0 / ( 1.0 + exp ( ( v - - 48.8 ) / 10.0 ) ) ;
+   hTau = ( 1.0 / ( hAlpha + hBeta ) ) ;
+     return 0; }
  
 #if 0 /*BBCORE*/
  
 static void _hoc_rates(void) {
   double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
   _thread = _extcall_thread;
   _nt = nrn_threads;
  _r = 1.;
- rates ( _threadargs_, *getarg(1) ;
+ rates ( _threadargs_ ;
  hoc_retpushx(_r);
 }
  
@@ -334,16 +364,26 @@ static void _hoc_rates(void) {
 
 static void initmodel(_threadargsproto_) {
   int _i; double _save;{
+  h = h0;
   m = m0;
  {
-   rates ( _threadargscomma_ v ) ;
-   m = minf ;
+   if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
+     mInf = 0.0 ;
+     mTau = 1e-12 ;
+     hInf = 1.0 ;
+     hTau = 1e-12 ;
+     }
+   else {
+     rates ( _threadargs_ ) ;
+     }
+   m = mInf ;
+   h = hInf ;
    }
  
 }
 }
 
-void nrn_init(_NrnThread* _nt, _Memb_list* _ml, int _type){
+void nrn_init(NrnThread* _nt, Memb_list* _ml, int _type){
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double _v, v; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -382,14 +422,17 @@ for (;;) { /* help clang-format properly indent */
     _PRCELLSTATE_V
  v = _v;
  _PRCELLSTATE_V
+  ena = _ion_ena;
+  ttxo = _ion_ttxo;
+  ttxi = _ion_ttxi;
  initmodel(_threadargs_);
  }
   }
 }
 
 static double _nrn_current(_threadargsproto_, double _v){double _current=0.;v=_v;{ {
-   gnasm = gnasmbar * m ;
-   ina = gnasm * ( v - ena ) ;
+   gNap_E = gNap_Ebar * m * m * m * h ;
+   ina = gNap_E * ( v - ena ) ;
    }
  _current += ina;
 
@@ -397,13 +440,13 @@ static double _nrn_current(_threadargsproto_, double _v){double _current=0.;v=_v
 }
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC)
-  void nrn_state_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_jacob_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_cur_launcher(_NrnThread*, _Memb_list*, int, int);
+  void nrn_state_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_jacob_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_cur_launcher(NrnThread*, Memb_list*, int, int);
 #endif
 
 
-void nrn_cur(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_cur(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 int* _ni; double _rhs, _g, _v, v; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -414,8 +457,8 @@ double * _vec_rhs = _nt->_actual_rhs;
 double * _vec_d = _nt->_actual_d;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_cur_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -439,6 +482,9 @@ for (;;) { /* help clang-format properly indent */
     int _nd_idx = _ni[_iml];
     _v = _vec_v[_nd_idx];
     _PRCELLSTATE_V
+  ena = _ion_ena;
+  ttxo = _ion_ttxo;
+  ttxi = _ion_ttxi;
  _g = _nrn_current(_threadargs_, _v + .001);
  	{ double _dina;
   _dina = ina;
@@ -455,7 +501,7 @@ for (;;) { /* help clang-format properly indent */
  
 }
 
-void nrn_state(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_state(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double v, _v = 0.0; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -464,8 +510,8 @@ _cntml_padded = _ml->_nodecount_padded;
 _thread = _ml->_thread;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_state_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -491,6 +537,9 @@ for (;;) { /* help clang-format properly indent */
     _PRCELLSTATE_V
  v=_v;
 {
+  ena = _ion_ena;
+  ttxo = _ion_ttxo;
+  ttxi = _ion_ttxi;
  {   states(_threadargs_);
   } }}
 
@@ -506,15 +555,13 @@ static void _initlists(){
  int _iml=0;
   if (!_first) return;
  
- _slist1 = (int*)malloc(sizeof(int)*1);
- _dlist1 = (int*)malloc(sizeof(int)*1);
+ _slist1 = (int*)malloc(sizeof(int)*2);
+ _dlist1 = (int*)malloc(sizeof(int)*2);
  _slist1[0] = &(m) - _p;  _dlist1[0] = &(Dm) - _p;
- #pragma acc enter data copyin(_slist1[0:1])
- #pragma acc enter data copyin(_dlist1[0:1])
+ _slist1[1] = &(h) - _p;  _dlist1[1] = &(Dh) - _p;
+ #pragma acc enter data copyin(_slist1[0:2])
+ #pragma acc enter data copyin(_dlist1[0:2])
 
 _first = 0;
 }
-
-#if defined(__cplusplus)
-} /* extern "C" */
-#endif
+} // namespace coreneuron_lib
