+++ test/validation/mod2c_core/cpp/Nap_No.cpp
@@ -15,14 +15,16 @@
 
 #include "coreneuron/scopmath_core/newton_struct.h"
 #include "coreneuron/nrnoc/md2redef.h"
+#include "coreneuron/nrnoc/register_mech.hpp"
 #if !NRNGPU
 #if !defined(DISABLE_HOC_EXP)
 #undef exp
 #define exp hoc_Exp
 #endif
 #endif
+ namespace coreneuron {
  
-#define _thread_present_ /**/ , _slist1[0:2], _dlist1[0:2] 
+#define _thread_present_ /**/ , _slist1[0:1], _dlist1[0:1] 
  
 #if defined(_OPENACC) && !defined(DISABLE_OPENACC)
 #include <openacc.h>
@@ -71,23 +73,23 @@
 #define _STRIDE _cntml_padded + _iml
 #endif
  
-#define nrn_init _nrn_init__Nap_E
-#define nrn_cur _nrn_cur__Nap_E
-#define _nrn_current _nrn_current__Nap_E
-#define nrn_jacob _nrn_jacob__Nap_E
-#define nrn_state _nrn_state__Nap_E
-#define initmodel initmodel__Nap_E
-#define _net_receive _net_receive__Nap_E
-#define nrn_state_launcher nrn_state_Nap_E_launcher
-#define nrn_cur_launcher nrn_cur_Nap_E_launcher
-#define nrn_jacob_launcher nrn_jacob_Nap_E_launcher 
-#define rates rates_Nap_E 
-#define states states_Nap_E 
+#define nrn_init _nrn_init__Nap_No
+#define nrn_cur _nrn_cur__Nap_No
+#define _nrn_current _nrn_current__Nap_No
+#define nrn_jacob _nrn_jacob__Nap_No
+#define nrn_state _nrn_state__Nap_No
+#define initmodel initmodel__Nap_No
+#define _net_receive _net_receive__Nap_No
+#define nrn_state_launcher nrn_state_Nap_No_launcher
+#define nrn_cur_launcher nrn_cur_Nap_No_launcher
+#define nrn_jacob_launcher nrn_jacob_Nap_No_launcher 
+#define rates rates_Nap_No 
+#define states states_Nap_No 
  
 #define _threadargscomma_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v,
-#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v,
+#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v,
 #define _threadargs_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v
-#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v
+#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v
  	/*SUPPRESS 761*/
 	/*SUPPRESS 762*/
 	/*SUPPRESS 763*/
@@ -96,26 +98,16 @@
  
 #define t _nt->_t
 #define dt _nt->_dt
-#define gNap_Ebar _p[0*_STRIDE]
-#define ina _p[1*_STRIDE]
-#define gNap_E _p[2*_STRIDE]
-#define m _p[3*_STRIDE]
-#define h _p[4*_STRIDE]
-#define ttxo _p[5*_STRIDE]
-#define ttxi _p[6*_STRIDE]
-#define ena _p[7*_STRIDE]
-#define mInf _p[8*_STRIDE]
-#define mTau _p[9*_STRIDE]
-#define mAlpha _p[10*_STRIDE]
-#define mBeta _p[11*_STRIDE]
-#define hInf _p[12*_STRIDE]
-#define hTau _p[13*_STRIDE]
-#define hAlpha _p[14*_STRIDE]
-#define hBeta _p[15*_STRIDE]
-#define Dm _p[16*_STRIDE]
-#define Dh _p[17*_STRIDE]
-#define _v_unused _p[18*_STRIDE]
-#define _g_unused _p[19*_STRIDE]
+#define gnamax _p[0*_STRIDE]
+#define gna _p[1*_STRIDE]
+#define pinf _p[2*_STRIDE]
+#define ptau _p[3*_STRIDE]
+#define p _p[4*_STRIDE]
+#define Dp _p[5*_STRIDE]
+#define ena _p[6*_STRIDE]
+#define ina _p[7*_STRIDE]
+#define _v_unused _p[8*_STRIDE]
+#define _g_unused _p[9*_STRIDE]
  
 #ifndef NRN_PRCELLSTATE
 #define NRN_PRCELLSTATE 0
@@ -130,8 +122,6 @@
 #define _ion_ena		_nt_data[_ppvar[0*_STRIDE]]
 #define _ion_ina	_nt_data[_ppvar[1*_STRIDE]]
 #define _ion_dinadv	_nt_data[_ppvar[2*_STRIDE]]
-#define _ion_ttxo		_nt_data[_ppvar[3*_STRIDE]]
-#define _ion_ttxi		_nt_data[_ppvar[4*_STRIDE]]
  
 #if MAC
 #if !defined(v)
@@ -141,10 +131,6 @@
 #define h _mlhh
 #endif
 #endif
- 
-#if defined(__cplusplus)
-extern "C" {
-#endif
  static int hoc_nrnpointerindex =  -1;
  static ThreadDatum* _extcall_thread;
  /* external NEURON variables */
@@ -152,6 +138,7 @@ extern "C" {
 #if 0 /*BBCORE*/
  /* declaration of user functions */
  static void _hoc_rates(void);
+ static void _hoc_vtrap(void);
  
 #endif /*BBCORE*/
  static int _mechtype;
@@ -159,12 +146,15 @@ extern "C" {
 #if 0 /*BBCORE*/
  /* connect user functions to hoc names */
  static VoidFunc hoc_intfunc[] = {
- "setdata_Nap_E", _hoc_setdata,
- "rates_Nap_E", _hoc_rates,
+ "setdata_Nap_No", _hoc_setdata,
+ "rates_Nap_No", _hoc_rates,
+ "vtrap_Nap_No", _hoc_vtrap,
  0, 0
 };
  
 #endif /*BBCORE*/
+#define vtrap vtrap_Nap_No
+ inline double vtrap( _threadargsprotocomma_ double , double );
  /* declare global and static user variables */
  
 static void _acc_globals_update() {
@@ -173,19 +163,19 @@ static void _acc_globals_update() {
 #if 0 /*BBCORE*/
  /* some parameters have upper and lower limits */
  static HocParmLimits _hoc_parm_limits[] = {
+ "gnamax_Nap_No", 0, 1e+09,
  0,0,0
 };
  static HocParmUnits _hoc_parm_units[] = {
- "gNap_Ebar_Nap_E", "S/cm2",
- "ina_Nap_E", "mA/cm2",
- "gNap_E_Nap_E", "S/cm2",
+ "gnamax_Nap_No", "S/cm2",
+ "gna_Nap_No", "S/cm2",
+ "ptau_Nap_No", "ms",
  0,0
 };
  
 #endif /*BBCORE*/
  static double delta_t = 0.01;
- static double h0 = 0;
- static double m0 = 0;
+ static double p0 = 0;
  /* connect global user variables to hoc */
  static DoubScal hoc_scdoub[] = {
  0,0
@@ -195,39 +185,34 @@ static void _acc_globals_update() {
 };
  static double _sav_indep;
  static void nrn_alloc(double*, Datum*, int);
-void nrn_init(_NrnThread*, _Memb_list*, int);
-void nrn_state(_NrnThread*, _Memb_list*, int);
- void nrn_cur(_NrnThread*, _Memb_list*, int);
+void nrn_init(NrnThread*, Memb_list*, int);
+void nrn_state(NrnThread*, Memb_list*, int);
+ void nrn_cur(NrnThread*, Memb_list*, int);
  /* connect range variables in _p that hoc is supposed to know about */
  static const char *_mechanism[] = {
  "6.2.0",
-"Nap_E",
- "gNap_Ebar_Nap_E",
+"Nap_No",
+ "gnamax_Nap_No",
  0,
- "ina_Nap_E",
- "gNap_E_Nap_E",
+ "gna_Nap_No",
+ "pinf_Nap_No",
+ "ptau_Nap_No",
  0,
- "m_Nap_E",
- "h_Nap_E",
+ "p_Nap_No",
  0,
  0};
  static int _na_type;
- static int _ttx_type;
  
 static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
  
 #if 0 /*BBCORE*/
  	/*initialize range parameters*/
- 	gNap_Ebar = 1e-05;
+ 	gnamax = 0.01;
  prop_ion = need_memb(_na_sym);
  nrn_promote(prop_ion, 0, 1);
  	_ppvar[0]._pval = &prop_ion->param[0]; /* ena */
  	_ppvar[1]._pval = &prop_ion->param[3]; /* ina */
  	_ppvar[2]._pval = &prop_ion->param[4]; /* _ion_dinadv */
- prop_ion = need_memb(_ttx_sym);
- nrn_promote(prop_ion, 1, 0);
- 	_ppvar[3]._pval = &prop_ion->param[2]; /* ttxo */
- 	_ppvar[4]._pval = &prop_ion->param[1]; /* ttxi */
  
 #endif /* BBCORE */
  
@@ -235,20 +220,18 @@ static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
  static void _initlists();
  static void _update_ion_pointer(Datum*);
  
-#define _psize 20
-#define _ppsize 5
- void _Nap_E_reg() {
+#define _psize 10
+#define _ppsize 3
+ void _Nap_No_reg() {
 	int _vectorized = 1;
   _initlists();
  _mechtype = nrn_get_mechtype(_mechanism[1]);
  if (_mechtype == -1) return;
  _nrn_layout_reg(_mechtype, LAYOUT);
- _na_type = nrn_get_mechtype("na_ion"); _ttx_type = nrn_get_mechtype("ttx_ion"); 
+ _na_type = nrn_get_mechtype("na_ion"); 
 #if 0 /*BBCORE*/
  	ion_reg("na", -10000.);
- 	ion_reg("ttx", 1.0);
  	_na_sym = hoc_lookup("na_ion");
- 	_ttx_sym = hoc_lookup("ttx_ion");
  
 #endif /*BBCORE*/
  	register_mech(_mechanism, nrn_alloc,nrn_cur, NULL, nrn_state, nrn_init, hoc_nrnpointerindex, 1);
@@ -256,107 +239,93 @@ static void nrn_alloc(double* _p, Datum* _ppvar, int _type) {
   hoc_register_dparam_semantics(_mechtype, 0, "na_ion");
   hoc_register_dparam_semantics(_mechtype, 1, "na_ion");
   hoc_register_dparam_semantics(_mechtype, 2, "na_ion");
-  hoc_register_dparam_semantics(_mechtype, 3, "ttx_ion");
-  hoc_register_dparam_semantics(_mechtype, 4, "ttx_ion");
  	hoc_register_var(hoc_scdoub, hoc_vdoub, NULL);
  }
-static char *modelname = "";
+static char *modelname = "Nap_No.mod   persistent sodium current";
 
 static int error;
 static int _ninits = 0;
 static int _match_recurse=1;
 static void _modl_cleanup(){ _match_recurse=1;}
-static int rates(_threadargsproto_);
+static int rates(_threadargsprotocomma_ double);
  
 static int _ode_spec1(_threadargsproto_);
 /*static int _ode_matsol1(_threadargsproto_);*/
  
-#define _slist1 _slist1_Nap_E
+#define _slist1 _slist1_Nap_No
 int* _slist1;
 #pragma acc declare create(_slist1)
 
-#define _dlist1 _dlist1_Nap_E
+#define _dlist1 _dlist1_Nap_No
 int* _dlist1;
 #pragma acc declare create(_dlist1)
  static inline int states(_threadargsproto_);
  
 /*CVODE*/
  static int _ode_spec1 (_threadargsproto_) {int _reset = 0; {
-   if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
-     mInf = 0.0 ;
-     mTau = 1e-12 ;
-     hInf = 1.0 ;
-     hTau = 1e-12 ;
-     }
-   else {
-     rates ( _threadargs_ ) ;
-     }
-   Dm = ( mInf - m ) / mTau ;
-   Dh = ( hInf - h ) / hTau ;
+   rates ( _threadargscomma_ v ) ;
+   Dp = ( pinf - p ) / ptau ;
    }
  return _reset;
 }
  static int _ode_matsol1 (_threadargsproto_) {
- if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
-   mInf = 0.0 ;
-   mTau = 1e-12 ;
-   hInf = 1.0 ;
-   hTau = 1e-12 ;
-   }
- else {
-   rates ( _threadargs_ ) ;
-   }
- Dm = Dm  / (1. - dt*( ( ( ( - 1.0 ) ) ) / mTau )) ;
- Dh = Dh  / (1. - dt*( ( ( ( - 1.0 ) ) ) / hTau )) ;
+ rates ( _threadargscomma_ v ) ;
+ Dp = Dp  / (1. - dt*( ( ( ( - 1.0 ) ) ) / ptau )) ;
  return 0;
 }
  /*END CVODE*/
  static int states (_threadargsproto_) { {
-   if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
-     mInf = 0.0 ;
-     mTau = 1e-12 ;
-     hInf = 1.0 ;
-     hTau = 1e-12 ;
-     }
-   else {
-     rates ( _threadargs_ ) ;
-     }
-    m = m + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / mTau)))*(- ( ( ( mInf ) ) / mTau ) / ( ( ( ( - 1.0) ) ) / mTau ) - m) ;
-    h = h + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / hTau)))*(- ( ( ( hInf ) ) / hTau ) / ( ( ( ( - 1.0) ) ) / hTau ) - h) ;
+   rates ( _threadargscomma_ v ) ;
+    p = p + (1. - exp(dt*(( ( ( - 1.0 ) ) ) / ptau)))*(- ( ( ( pinf ) ) / ptau ) / ( ( ( ( - 1.0) ) ) / ptau ) - p) ;
    }
   return 0;
 }
  
-static int  rates ( _threadargsproto_ ) {
-    mInf = 1.0 / ( 1.0 + exp ( ( v - - 52.6 ) / - 4.6 ) ) ;
-   if ( v  == - 25.0 ) {
-     v = v + 0.0001 ;
-     }
-   mAlpha = ( 0.182 * ( v - - 25.0 ) ) / ( 1.0 - ( exp ( - ( v - - 25.0 ) / 9.0 ) ) ) ;
-   mBeta = ( 0.124 * ( - v - 25.0 ) ) / ( 1.0 - ( exp ( - ( - v - 25.0 ) / 9.0 ) ) ) ;
-   mTau = 6.0 * 1.0 / ( mAlpha + mBeta ) ;
-   if ( v  == - 50.0 ) {
-     v = v + 0.0001 ;
-     }
-   hAlpha = ( 0.00001 * ( v - - 50.0 ) ) / ( 1.0 - ( exp ( - ( v - - 50.0 ) / 13.0 ) ) ) ;
-   if ( v  == - 75.0 ) {
-     v = v + 0.0001 ;
-     }
-   hBeta = ( 0.00001 * ( - v - 75.0 ) ) / ( 1.0 - ( exp ( - ( - v - 75.0 ) / 13.0 ) ) ) ;
-   hInf = 1.0 / ( 1.0 + exp ( ( v - - 48.8 ) / 10.0 ) ) ;
-   hTau = ( 1.0 / ( hAlpha + hBeta ) ) ;
-     return 0; }
+static int  rates ( _threadargsprotocomma_ double _lv ) {
+   double _lalpha , _lbeta , _lsum ;
+  _lalpha = 0.0353 * vtrap ( _threadargscomma_ - 17.0 - _lv , 10.2 ) ;
+   _lbeta = 0.000883 * vtrap ( _threadargscomma_ _lv + 24.0 , 10.0 ) ;
+   _lsum = _lalpha + _lbeta ;
+   ptau = 1.0 / _lsum ;
+   pinf = _lalpha / _lsum ;
+    return 0; }
  
 #if 0 /*BBCORE*/
  
 static void _hoc_rates(void) {
   double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
   _thread = _extcall_thread;
   _nt = nrn_threads;
  _r = 1.;
- rates ( _threadargs_ ;
+ rates ( _threadargs_, *getarg(1) ;
+ hoc_retpushx(_r);
+}
+ 
+#endif /*BBCORE*/
+ 
+double vtrap ( _threadargsprotocomma_ double _lx , double _ly ) {
+   double _lvtrap;
+ if ( fabs ( _lx / _ly ) < 1e-6 ) {
+     _lvtrap = _ly * ( 1.0 - _lx / _ly / 2.0 ) ;
+     }
+   else {
+     _lvtrap = _lx / ( exp ( _lx / _ly ) - 1.0 ) ;
+     }
+   
+return _lvtrap;
+ }
+ 
+#if 0 /*BBCORE*/
+ 
+static void _hoc_vtrap(void) {
+  double _r;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
+   if (_extcall_prop) {_p = _extcall_prop->param; _ppvar = _extcall_prop->dparam;}else{ _p = (double*)0; _ppvar = (Datum*)0; }
+  _thread = _extcall_thread;
+  _nt = nrn_threads;
+ _r =  vtrap ( _threadargs_, *getarg(1) , *getarg(2) ;
  hoc_retpushx(_r);
 }
  
@@ -366,26 +335,16 @@ static void _hoc_rates(void) {
 
 static void initmodel(_threadargsproto_) {
   int _i; double _save;{
-  h = h0;
-  m = m0;
+  p = p0;
  {
-   if ( ttxi  == 0.015625  && ttxo > 1e-12 ) {
-     mInf = 0.0 ;
-     mTau = 1e-12 ;
-     hInf = 1.0 ;
-     hTau = 1e-12 ;
-     }
-   else {
-     rates ( _threadargs_ ) ;
-     }
-   m = mInf ;
-   h = hInf ;
+   rates ( _threadargscomma_ v ) ;
+   p = pinf ;
    }
  
 }
 }
 
-void nrn_init(_NrnThread* _nt, _Memb_list* _ml, int _type){
+void nrn_init(NrnThread* _nt, Memb_list* _ml, int _type){
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double _v, v; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -425,16 +384,14 @@ for (;;) { /* help clang-format properly indent */
  v = _v;
  _PRCELLSTATE_V
   ena = _ion_ena;
-  ttxo = _ion_ttxo;
-  ttxi = _ion_ttxi;
  initmodel(_threadargs_);
  }
   }
 }
 
 static double _nrn_current(_threadargsproto_, double _v){double _current=0.;v=_v;{ {
-   gNap_E = gNap_Ebar * m * m * m * h ;
-   ina = gNap_E * ( v - ena ) ;
+   gna = gnamax * p * p * p ;
+   ina = gna * ( v - ena ) ;
    }
  _current += ina;
 
@@ -442,13 +399,13 @@ static double _nrn_current(_threadargsproto_, double _v){double _current=0.;v=_v
 }
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC)
-  void nrn_state_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_jacob_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_cur_launcher(_NrnThread*, _Memb_list*, int, int);
+  void nrn_state_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_jacob_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_cur_launcher(NrnThread*, Memb_list*, int, int);
 #endif
 
 
-void nrn_cur(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_cur(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 int* _ni; double _rhs, _g, _v, v; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -459,8 +416,8 @@ double * _vec_rhs = _nt->_actual_rhs;
 double * _vec_d = _nt->_actual_d;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_cur_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -485,8 +442,6 @@ for (;;) { /* help clang-format properly indent */
     _v = _vec_v[_nd_idx];
     _PRCELLSTATE_V
   ena = _ion_ena;
-  ttxo = _ion_ttxo;
-  ttxi = _ion_ttxi;
  _g = _nrn_current(_threadargs_, _v + .001);
  	{ double _dina;
   _dina = ina;
@@ -503,7 +458,7 @@ for (;;) { /* help clang-format properly indent */
  
 }
 
-void nrn_state(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_state(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double v, _v = 0.0; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -512,8 +467,8 @@ _cntml_padded = _ml->_nodecount_padded;
 _thread = _ml->_thread;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_state_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -540,8 +495,6 @@ for (;;) { /* help clang-format properly indent */
  v=_v;
 {
   ena = _ion_ena;
-  ttxo = _ion_ttxo;
-  ttxi = _ion_ttxi;
  {   states(_threadargs_);
   } }}
 
@@ -557,16 +510,12 @@ static void _initlists(){
  int _iml=0;
   if (!_first) return;
  
- _slist1 = (int*)malloc(sizeof(int)*2);
- _dlist1 = (int*)malloc(sizeof(int)*2);
- _slist1[0] = &(m) - _p;  _dlist1[0] = &(Dm) - _p;
- _slist1[1] = &(h) - _p;  _dlist1[1] = &(Dh) - _p;
- #pragma acc enter data copyin(_slist1[0:2])
- #pragma acc enter data copyin(_dlist1[0:2])
+ _slist1 = (int*)malloc(sizeof(int)*1);
+ _dlist1 = (int*)malloc(sizeof(int)*1);
+ _slist1[0] = &(p) - _p;  _dlist1[0] = &(Dp) - _p;
+ #pragma acc enter data copyin(_slist1[0:1])
+ #pragma acc enter data copyin(_dlist1[0:1])
 
 _first = 0;
 }
-
-#if defined(__cplusplus)
-} /* extern "C" */
-#endif
+} // namespace coreneuron_lib
