+++ test/validation/mod2c_core/cpp/SynNMDA10_2_2.cpp
@@ -15,12 +15,14 @@
 
 #include "coreneuron/scopmath_core/newton_struct.h"
 #include "coreneuron/nrnoc/md2redef.h"
+#include "coreneuron/nrnoc/register_mech.hpp"
 #if !NRNGPU
 #if !defined(DISABLE_HOC_EXP)
 #undef exp
 #define exp hoc_Exp
 #endif
 #endif
+ namespace coreneuron {
  
 #define _thread_present_ /**/ , _thread[0:2] , _slist1[0:10], _dlist1[0:10] 
  
@@ -88,7 +90,7 @@
 #define nrn_jacob_launcher nrn_jacob_NMDA10_2_2_launcher 
 #if NET_RECEIVE_BUFFERING
 #define _net_buf_receive _net_buf_receive_NMDA10_2_2
-static void _net_buf_receive(_NrnThread*);
+static void _net_buf_receive(NrnThread*);
 #endif
  
 #define kstates kstates_NMDA10_2_2 
@@ -96,9 +98,9 @@ static void _net_buf_receive(_NrnThread*);
 #define rates rates_NMDA10_2_2 
  
 #define _threadargscomma_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v,
-#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v,
+#define _threadargsprotocomma_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v,
 #define _threadargs_ _iml, _cntml_padded, _p, _ppvar, _thread, _nt, v
-#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, _NrnThread* _nt, double v
+#define _threadargsproto_ int _iml, int _cntml_padded, double* _p, Datum* _ppvar, ThreadDatum* _thread, NrnThread* _nt, double v
  	/*SUPPRESS 761*/
 	/*SUPPRESS 762*/
 	/*SUPPRESS 763*/
@@ -169,10 +171,6 @@ static void _net_buf_receive(_NrnThread*);
 #if !defined(h)
 #define h _mlhh
 #endif
-#endif
- 
-#if defined(__cplusplus)
-extern "C" {
 #endif
  static int hoc_nrnpointerindex =  -1;
  static ThreadDatum* _extcall_thread;
@@ -281,9 +279,9 @@ static void _acc_globals_update() {
 };
  static double _sav_indep;
  static void nrn_alloc(double*, Datum*, int);
-void nrn_init(_NrnThread*, _Memb_list*, int);
-void nrn_state(_NrnThread*, _Memb_list*, int);
- void nrn_cur(_NrnThread*, _Memb_list*, int);
+void nrn_init(NrnThread*, Memb_list*, int);
+void nrn_state(NrnThread*, Memb_list*, int);
+ void nrn_cur(NrnThread*, Memb_list*, int);
  
 #if 0 /*BBCORE*/
  static void _hoc_destroy_pnt(_vptr) void* _vptr; {
@@ -581,9 +579,9 @@ for(_i=1;_i<10;_i++){
 #undef t
 #define t _nrb_t
 static void _net_receive_kernel(double, Point_process*, int _weight_index, double _flag);
-static void _net_buf_receive(_NrnThread* _nt) {
+static void _net_buf_receive(NrnThread* _nt) {
   if (!_nt->_ml_list) { return; }
-  _Memb_list* _ml = _nt->_ml_list[_mechtype];
+  Memb_list* _ml = _nt->_ml_list[_mechtype];
   if (!_ml) { return; }
   NetReceiveBuffer_t* _nrb = _ml->_net_receive_buffer; 
   int _di;
@@ -646,7 +644,7 @@ static void _net_send_buffering(NetSendBuffer_t* _nsb, int _sendtype, int _i_vda
 }
  
 static void _net_receive (Point_process* _pnt, int _weight_index, double _lflag) {
-  _NrnThread* _nt = nrn_threads + _pnt->_tid;
+  NrnThread* _nt = nrn_threads + _pnt->_tid;
   NetReceiveBuffer_t* _nrb = _nt->_ml_list[_mechtype]->_net_receive_buffer;
   if (_nrb->_cnt >= _nrb->_size){
     realloc_net_receive_buffer(_nt, _nt->_ml_list[_mechtype]);
@@ -665,9 +663,9 @@ static void _net_receive (Point_process* _pnt, int _weight_index, double _lflag)
 #endif
  
 {  double* _p; Datum* _ppvar; ThreadDatum* _thread; double v;
-   _Memb_list* _ml; int _cntml_padded, _cntml_actual; int _iml; double* _args;
+   Memb_list* _ml; int _cntml_padded, _cntml_actual; int _iml; double* _args;
  
-   _NrnThread* _nt;
+   NrnThread* _nt;
    int _tid = _pnt->_tid; 
    _nt = nrn_threads + _tid;
    _thread = (ThreadDatum*)0; 
@@ -726,11 +724,11 @@ static int  release ( _threadargsprotocomma_ double _lt ) {
  
 static double _hoc_release(void* _vptr) {
  double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    _p = ((Point_process*)_vptr)->_prop->param;
   _ppvar = ((Point_process*)_vptr)->_prop->dparam;
   _thread = _extcall_thread;
-  _nt = (_NrnThread*)((Point_process*)_vptr)->_vnt;
+  _nt = (NrnThread*)((Point_process*)_vptr)->_vnt;
  _r = 1.;
  release ( _threadargs_, *getarg(1) );
  return(_r);
@@ -747,11 +745,11 @@ static int  rates ( _threadargsprotocomma_ double _lv ) {
  
 static double _hoc_rates(void* _vptr) {
  double _r;
-   double* _p; Datum* _ppvar; ThreadDatum* _thread; _NrnThread* _nt;
+   double* _p; Datum* _ppvar; ThreadDatum* _thread; NrnThread* _nt;
    _p = ((Point_process*)_vptr)->_prop->param;
   _ppvar = ((Point_process*)_vptr)->_prop->dparam;
   _thread = _extcall_thread;
-  _nt = (_NrnThread*)((Point_process*)_vptr)->_vnt;
+  _nt = (NrnThread*)((Point_process*)_vptr)->_vnt;
  _r = 1.;
  rates ( _threadargs_, *getarg(1) );
  return(_r);
@@ -933,7 +931,7 @@ static void _thread_cleanup(ThreadDatum* _thread) {
 
 static void initmodel(_threadargsproto_) {
   int _i; double _save;{
-  _Memb_list* _ml = _nt->_ml_list[_mechtype];
+  Memb_list* _ml = _nt->_ml_list[_mechtype];
   CB2 = CB20;
   CB1 = CB10;
   CB0 = CB00;
@@ -961,7 +959,7 @@ static void initmodel(_threadargsproto_) {
    OB = 0.0 ;
    
 #if NET_RECEIVE_BUFFERING
-    _net_send_buffering(_ml->_net_send_buffer, 0, _tqitem, 0, _ppvar[1*_STRIDE], t +  590.0 , 1.0 );
+    _net_send_buffering(_ml->_net_send_buffer, 0, _tqitem, 0,  _ppvar[1*_STRIDE], t +  590.0 , 1.0 );
 #else
  net_send ( _tqitem, -1, (Point_process*) _nt->_vdata[_ppvar[1*_STRIDE]], t +  590.0 , 1.0 ) ;
    
@@ -971,7 +969,7 @@ static void initmodel(_threadargsproto_) {
 }
 }
 
-void nrn_init(_NrnThread* _nt, _Memb_list* _ml, int _type){
+void nrn_init(NrnThread* _nt, Memb_list* _ml, int _type){
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double _v, v; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -1056,13 +1054,13 @@ static double _nrn_current(_threadargsproto_, double _v){double _current=0.;v=_v
 }
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC)
-  void nrn_state_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_jacob_launcher(_NrnThread*, _Memb_list*, int, int);
-  void nrn_cur_launcher(_NrnThread*, _Memb_list*, int, int);
+  void nrn_state_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_jacob_launcher(NrnThread*, Memb_list*, int, int);
+  void nrn_cur_launcher(NrnThread*, Memb_list*, int, int);
 #endif
 
 
-void nrn_cur(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_cur(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 int* _ni; double _rhs, _g, _v, v; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -1075,8 +1073,8 @@ double * _vec_shadow_rhs = _nt->_shadow_rhs;
 double * _vec_shadow_d = _nt->_shadow_d;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_cur_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -1143,7 +1141,7 @@ for (;;) { /* help clang-format properly indent */
  
 }
 
-void nrn_state(_NrnThread* _nt, _Memb_list* _ml, int _type) {
+void nrn_state(NrnThread* _nt, Memb_list* _ml, int _type) {
 double* _p; Datum* _ppvar; ThreadDatum* _thread;
 double v, _v = 0.0; int* _ni; int _iml, _cntml_padded, _cntml_actual;
     _ni = _ml->_nodeindices;
@@ -1152,8 +1150,8 @@ _cntml_padded = _ml->_nodecount_padded;
 _thread = _ml->_thread;
 
 #if defined(ENABLE_CUDA_INTERFACE) && defined(_OPENACC) && !defined(DISABLE_OPENACC)
-  _NrnThread* d_nt = acc_deviceptr(_nt);
-  _Memb_list* d_ml = acc_deviceptr(_ml);
+  NrnThread* d_nt = acc_deviceptr(_nt);
+  Memb_list* d_ml = acc_deviceptr(_ml);
   nrn_state_launcher(d_nt, d_ml, _type, _cntml_actual);
   return;
 #endif
@@ -1183,7 +1181,7 @@ for (;;) { /* help clang-format properly indent */
   #if !defined(_kinetic_kstates_NMDA10_2_2)
     #define _kinetic_kstates_NMDA10_2_2 0
   #endif
-  sparse_thread(_thread[_spth1]._pvoid, 10, _slist1, _dlist1, &t, dt, _kinetic_kstates_NMDA10_2_2, _linmat1, _threadargs_);
+  sparse_thread((SparseObj*)_thread[_spth1]._pvoid, 10, _slist1, _dlist1, &t, dt, _kinetic_kstates_NMDA10_2_2, _linmat1, _threadargs_);
   }}}
 
 }
@@ -1215,7 +1213,4 @@ static void _initlists(){
 
 _first = 0;
 }
-
-#if defined(__cplusplus)
-} /* extern "C" */
-#endif
+} // namespace coreneuron_lib
